+++ game_patch/rf/geometry.h
@@ -130,6 +130,11 @@ namespace rf
         {
             AddrCaller{0x004DF1C0}.this_call(this, in, out, clear_fraction);
         }
+
+        void set_levelmod_blast_autotexture_ppm(float ppm)
+        {
+            AddrCaller{0x004F8730}.this_call(this, ppm);
+        }
     };
     static_assert(sizeof(GSolid) == 0x378);